a nice state
