+++ libs/cy-reporter/src/list.ts
@@ -1,8 +1,13 @@
 import './env'
-import { init } from './sender'
+import { init, db } from './data-source'
+import { Repository } from 'typeorm'
+import { CypressRun } from './entity'
 
-init.then((repo) => {
-  repo.find().then((res) => {
-    console.log(res)
+const main = async (repo: Repository<CypressRun>) => {
+  return repo.find().then((res) => {
+    console.log(res, res.length)
   })
-})
+}
+
+// run main under an initialized connection
+init.then((repo) => main(repo)).then(() => db.destroy())
